modified:   sudo-taijutsu.sh
In testing, I found that an Alias I wasn't expecting to be deleted, was
being deleted. It happened so fast back to back to the user in question,
that in the user+host+group list I compiled, I did not realize a
near-adjacent account name was responsible for the Alias' haven't been
deleted.

It wasn't until I added some debug code with a line count, that the
issue was actually revealed:

    if [[ "${curUsername,,}" == "foo" ]] ;
    then
      echo -e "\n\n\n\n\t\t\t Current user is foo!!!\n\n\n";
      echo "is their aslias still there?"
      grep -E "Alias.*${curUsername}" ${fileSudoers}
      intBefore=$(wc -l "${fileSudoers}" | awk '{print $1}')
      sleep 10;
    fi

    ${cmdEcho} -e "\n${strStep}\n" | ${cmdTee} "${fileLog}"
    sed -i -E "/^[^=]*\b${curUsername}\b[^=]*=/Id" "${fileSudoers}";

    if [[ "${curUsername}" == "foo" ]] ;
    then
      echo -e "\n\n\n\n\t\t\t Current user foo was just deleted!!! \n\n\n";
      intAfter=$(wc -l "${fileSudoers}" | awk '{print $1}')
      echo "For step ${curUsername}, $(echo ${intBefore} - ${intAfter} | bc) lines were deleted."
      sleep 60;
      read -n 1 -s -r -p "Press any key to continue"
    fi

(username replaced by "foo" for this checkin)

I expected 686 lines to be deleted, but only 685 lines were deleted. The
cuprit was a preceding groupname, for a user Alias definition.

However, because these were so close, I discovered there was a bug which
caused the regex to be overly-greedy, which resulted in the deletion of
3,000 lines that should not have been deleted, so in chasing down this
false bug, I actually solved a bigger bug several days before I would
have.

I'm going to do another commit back to back with this removing the above
debug code, but wanted to commit this with the notes for proper record
keeping for future maintenance (in event we ever introduce a regression
bug).