modified:   README.md
modified:   sudo-taijutsu.sh

Added a start timestamp to the log (moved it past command processing and
function definitions so the logging alias would work)

Modified the custom filter, have more testing to do. This is where your
tribal knowledge is key. I do not know which User_Alias items that are
not in the Active Directory or cmdb must be preserved, so I am making
educated guesses. This is why your staff need to be testing this. I can
test raw functionality but not 100% usability in your particular
environment.