renamed:    sudo-katana.sh -> sudo-chop.sh renamed:    sudo-taijutsu.sh -> sudo-cleanup.sh
Since this is entering production, we're changing the name from the
working names to the final utility names. Documentation updates
coming.

Also, the evaluation branch has been deleted in favor of "unstable"
and work will commence in unstable, with main intended to be stable.

After documentation is updated, release branches will be create for
each release.